Fix CSScomb with Sass tmLanguage
The Sass tmLanguage package can be used to highlight SCSS and Sass files. If
used on an SCSS file, CSSComb does not run (since it does not match the scope
name as starting with `source.scss`).

This commit enables CSSComb for both Sass and SCSS files when being highlighted
with the Sass tmLanguage package.

Fixes #18